Remove community posts poc
